Refactor day 09 to use walrus and functools.reduce
$ hyperfine 'python 09.py <data/09.txt'
Benchmark 1: python 09.py <data/09.txt
  Time (mean ± σ):      24.8 ms ±   0.5 ms    [User: 20.7 ms, System: 4.1 ms]
  Range (min … max):    24.0 ms …  26.4 ms    105 runs

Importing functools adds about ~3ms of runtime -- which is more than 10%
of this solution's total. That's a lot to pay for saving a couple lines
of code... 😂

NOTE: I realized that the main `sum` stopping condition is bugged, e.g.
diff = [5, -5, 5, -5]. It works for both the examples and my input so
I'm keeping it. \shrug